+++ .idea/consolestorage/flexible_search_region.json
@@ -1,0 +1,5 @@
+{
+  "name" : "FLEXIBLE_SEARCH",
+  "entities" : { },
+  "maxNumberEntities" : 7
+}
