+++ .idea/consolestorage/solr_region.json
@@ -1,0 +1,5 @@
+{
+  "name" : "SOLR",
+  "entities" : { },
+  "maxNumberEntities" : 7
+}
