+++ spring-boot/src/main/java/com/springbackend/app/cleandata/DataCleanupTask.java
@@ -1,22 +1,20 @@
 package com.springbackend.app.cleandata;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import com.springbackend.app.dao.EarthquakeDao;
-import com.springbackend.app.model.Earthquake;
-import java.util.List;
 import com.springbackend.app.service.EarthquakeService;
 
 
 @Component
+@RequiredArgsConstructor
 public class DataCleanupTask {
 
-    @Autowired
-	EarthquakeService earthService;
+	private final EarthquakeService earthService;
 
     @Scheduled(fixedRate = 10000)
     public void cleanupData() {
-    earthService.cleanEarthquakeData();
+        earthService.cleanEarthquakeData();
     }
 }
