--- spring-boot/src/main/java/com/springbackend/app/dao/EarthquakeDao.java
@@ -1,11 +1,0 @@
-package com.springbackend.app.dao;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.springbackend.app.model.Earthquake;
-
-@Repository
-public interface EarthquakeDao extends JpaRepository<Earthquake, Long> {
-
-}
