+++ spring-boot/src/main/resources/application.properties
@@ -10,8 +10,13 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/h2
 spring.h2.console.settings.trace=false
 
-spring.datasource.data=*.sql
-spring.datasource.initialization-mode=always
+spring.sql.init.data-locations=*.sql
+spring.sql.init.mode=always
 spring.jpa.hibernate.ddl-auto=create-drop
 
 server.error.include-stacktrace=never
+
+logging.file.name=logs/KartacaProject.log
+
+logging.level.root=INFO
+
