+++ spring-boot/src/main/resources/import.sql
@@ -1,3 +1,3 @@
-INSERT INTO TBL_EARTHQUAKE (id,lat, lon, magnitude) VALUES (34, 89, 5.6);
-INSERT INTO TBL_EARTHQUAKE (id,lat, lon, magnitude) VALUES (89, 23, 2.6);
-INSERT INTO TBL_EARTHQUAKE (id,lat, lon, magnitude) VALUES (98, 23, 5.9);
+INSERT INTO TBL_EARTHQUAKE (lat, lon, magnitude) VALUES (89, 5.6, 34);
+INSERT INTO TBL_EARTHQUAKE (lat, lon, magnitude) VALUES (23, 2.6, 89);
+INSERT INTO TBL_EARTHQUAKE (lat, lon, magnitude) VALUES (23, 5.9, 98);
